+++ src/functions/implements/activation/selu.c
@@ -41,8 +41,8 @@ rt_function_error_t exec_selu(rt_function_t *f) {
   const int size = calc_shape_size(f->inputs[0]->shape);
   int s;
   for (s = 0; s < size; s++) {
-    y[s] =
-        x[s] > (float)0 ? context->scale * x[s] : coef * (exp(x[s]) - (float)1);
+    y[s] = (float)(x[s] > (float)0 ? context->scale * x[s]
+                                   : coef * (exp(x[s]) - (float)1));
   }
   return RT_FUNCTION_ERROR_NOERROR;
 }
@@ -58,8 +58,8 @@ rt_function_error_t exec_selu_generic(rt_function_t *f) {
   int s;
   for (s = 0; s < size; s++) {
     float val_x = get_input(input, s);
-    float val_y = val_x > (float)0 ? context->scale * val_x
-                                   : coef * (exp(val_x) - (float)1);
+    float val_y = (float)(val_x > (float)0 ? context->scale * val_x
+                                           : coef * (exp(val_x) - (float)1));
     set_output(output, s, val_y);
   }
   return RT_FUNCTION_ERROR_NOERROR;
