+++ src/functions/implements/math/round.c
@@ -75,7 +75,7 @@ rt_function_error_t exec_round(rt_function_t *f) {
 
   int i; // Iterator
   for (i = 0; i < p->output_size; i++) {
-    y[i] = round(x[i]);
+    y[i] = (float)round(x[i]);
   }
   return RT_FUNCTION_ERROR_NOERROR;
 }
